backend/requirements.txt
alembic==1.16.1
blinker==1.8.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.10
MarkupSafe==3.0.2
PyMySQL==1.1.1
python-dotenv==1.1.0
requests==2.32.4
SQLAlchemy==2.0.36
stripe==12.2.0
typing_extensions==4.12.2
urllib3==2.4.0
Werkzeug==3.0.6
